+++ lang/en/admin.php
@@ -443,7 +443,7 @@ return [
             'ldap' => 'LDAP',
             'local' => 'Local',
             'shibboleth' => 'Shibboleth',
-            'title' => 'Authentication Type',
+            'title' => 'Login method',
         ],
         'base_data' => 'Base data',
         'bbb' => 'BigBlueButton',