+++ tests/Frontend/e2e/UserProfileBase.cy.js
@@ -42,7 +42,7 @@ describe("User Profile Base", function () {
     // Check authenticator setting
     cy.get('[data-test="authenticator-field"]')
       .should("be.visible")
-      .and("include.text", "auth.authenticator")
+      .and("include.text", "admin.users.authenticator.title")
       .within(() => {
         cy.get("#authenticator")
           .should("have.value", "admin.users.authenticator.local")
@@ -705,10 +705,10 @@ describe("User Profile Base", function () {
       .should("have.value", "admin.users.authenticator.ldap")
       .and("be.disabled");
 
-    cy.get('[data-test="authenticator-id-field"]')
-      .should("include.text", "auth.authenticator_id")
+    cy.get('[data-test="external-user-id-field"]')
+      .should("include.text", "auth.external_user_id")
       .within(() => {
-        cy.get("#authenticator_id")
+        cy.get("#external-user-id")
           .should("have.value", "jdo")
           .and("be.disabled");
       });